Move timeline drawer into it's own directory
