+++ packages/studio-base/src/components/TimelineDrawer/TimelineScrubber.tsx
